Generate test data and modify directory structure.
For the issue #54
