+++ README.md
@@ -102,3 +102,38 @@ Sean 的保送之路
 1. 看完「培養與鍛鍊程式設計的邏輯腦」Page 42 - 47
 2. 並實作「小數點字串換算浮點數」，類遞推法題目
 
+## 20151115
+1. 寫 closet pair
+2. 完成 a058 MOD3
+
+##### 回家作業：
+1. 實作「小數點字串換算浮點數」，類遞推法題目
+
+## 20151121
+1. 完成 a147 print it all
+
+##### 回家作業：
+1. 實作「小數點字串換算浮點數」，類遞推法題目
+
+## 20151129
+1. 講解 「培養與鍛鍊程式設計的邏輯腦」ch5
+2. 講解輾轉相除法 & 實作
+3. 演算法概念：Divide and Conquer
+4. Recursive: Fraction, Summation
+
+##### 回家作業：
+1. 看懂 merge sort
+
+## 20151205
+1. 完成 a059 完全平方和
+
+##### 回家作業：
+1. 研究二分搜尋法
+
+## 20151205
+1. 練習手寫解二分搜尋法
+2. 實作二分搜尋法
+
+##### 回家作業：
+1. 完成二分搜尋法
+2. 完成 a149 乘乘樂
