+++ algorithm_practice/resources.c
@@ -1,0 +1,6 @@
+#include <stdio.h>
+
+int main(int argc, char const *argv[]) {
+	
+	return 0;
+}
