--- resources.c
@@ -1,6 +1,0 @@
-#include <stdio.h>
-
-int main(int argc, char const *argv[]) {
-	
-	return 0;
-}
