+++ test_data/a.in
@@ -1,0 +1,100 @@
+1 16808 475250
+2 650074 943659
+3 108931 211273
+4 27545 850879
+5 777924 237710
+6 564441 438166
+7 484493 243043
+8 807988 522504
+9 282328 531730
+10 378841 542613
+11 544304 833170
+12 817710 97158
+13 129561 570934
+14 493100 280279
+15 351817 795336
+16 899098 807827
+17 613513 929268
+18 723811 877634
+19 580980 379150
+20 236580 558822
+21 511968 10673
+22 901394 819337
+23 125486 91746
+24 425229 894092
+25 640195 186358
+26 35002 921154
+27 616709 357945
+28 515669 571491
+29 788125 702197
+30 509531 60904
+31 727723 454667
+32 728550 118025
+33 797802 806854
+34 540978 847409
+35 308229 374934
+36 260299 938982
+37 28636 408014
+38 623866 389815
+39 739064 554537
+40 339426 1670
+41 724116 220095
+42 75630 446502
+43 546518 934196
+44 30106 810405
+45 319452 194299
+46 782189 821124
+47 799506 246883
+48 406753 731567
+49 936717 130338
+50 354439 53145
+51 851502 844898
+52 783872 955829
+53 111138 619359
+54 153178 982398
+55 722295 477905
+56 621610 946232
+57 959746 844176
+58 853636 266299
+59 98143 880400
+60 827969 90413
+61 400261 673558
+62 441595 688009
+63 716396 325969
+64 119114 204531
+65 471007 910963
+66 797943 285366
+67 842083 179853
+68 530768 633822
+69 185696 824713
+70 973672 859902
+71 248591 724832
+72 314739 153058
+73 631617 819791
+74 555641 110680
+75 437336 77007
+76 497973 936099
+77 649096 975320
+78 583455 345224
+79 112290 556761
+80 948906 586127
+81 236124 129507
+82 45814 476771
+83 935239 329095
+84 896221 987845
+85 712367 140535
+86 725227 915395
+87 941364 318739
+88 607845 559591
+89 436551 145160
+90 808624 764948
+91 597386 307218
+92 195273 755540
+93 399248 242386
+94 413497 894886
+95 140624 682421
+96 28145 531969
+97 335736 699916
+98 161626 663535
+99 46043 796012
+100 454680 24153
