+++ test_data/bs_data1.in
@@ -1,0 +1,21 @@
+10
+11
+4 5 5 12 16 16 17 20 42 45 54 56 63 66 70 73 74 79 85 95 95
+55
+6 11 14 21 31 31 46 51 69 69 78 89 96 98
+92
+13 13 15 30 30 32 35 40 41 43 46 58 60 74 84 88
+84
+0 3 4 21 25 27 32 36 36 37 46 51 57 61 65 68 71 75 78 79 83 89
+100
+6 14 24 35 39 47 54 60 63 78 88 88 90 93
+24
+0 2 5 11 18 19 19 22 24 30 33 37 38 49 50 66 74 76 81 83 86 87 90 95
+96
+11 12 17 19 25 28 29 31 31 36 37 40 46 48 50 50 55 57 59 61 69 72 81 82 85 90
+26
+16 29 32 33 37 46 50 56 61 87 88
+18
+0 11 12 12 13 17 36 38 38 39 44 52 53 55 57 66 69 72 81 87 90
+2
+0 4 16 27 30 38 40 47 50 51 51 53 56 64 66 71 80 83 87 90 99
