+++ test_data/e.in
@@ -1,0 +1,5 @@
+1 2 7
+2 4 6
+3 5 1
+4 8 2
+5 2 1
