+++ test_data/float_string1.log
@@ -1,0 +1,100 @@
+8.75250
+74.43659
+31.11273
+45.50879
+24.37710
+41.38166
+93.43043
+88.22504
+28.31730
+41.42613
+4.33170
+10.97158
+61.70934
+100.80279
+17.95336
+98.7827
+13.29268
+11.77634
+80.79150
+80.58822
+68.10673
+94.19337
+86.91746
+29.94092
+95.86358
+2.21154
+9.57945
+69.71491
+25.2197
+31.60904
+23.54667
+50.18025
+2.6854
+78.47409
+29.74934
+99.38982
+36.8014
+66.89815
+64.54537
+26.1670
+16.20095
+30.46502
+18.34196
+6.10405
+52.94299
+89.21124
+6.46883
+53.31567
+17.30338
+39.53145
+2.44898
+72.55829
+38.19359
+78.82398
+95.77905
+10.46232
+46.44176
+36.66299
+43.80400
+69.90413
+61.73558
+95.88009
+96.25969
+14.4531
+7.10963
+43.85366
+83.79853
+68.33822
+96.24713
+72.59902
+91.24832
+39.53058
+17.19791
+41.10680
+36.77007
+73.36099
+96.75320
+55.45224
+90.56761
+6.86127
+24.29507
+14.76771
+39.29095
+21.87845
+67.40535
+27.15395
+64.18739
+45.59591
+51.45160
+24.64948
+86.7218
+73.55540
+48.42386
+97.94886
+24.82421
+45.31969
+36.99916
+26.63535
+43.96012
+80.24153
